Added initial memory management for required metadata
Now the handle used by the superset disassembler wraps Capstone's
handle, but also stores some additional data of its own.

Unfortunately it can't use the same memory allocation functions
as Capstone, perhaps because they aren't exported.  Strangely, the
symbols are present, but when trying to call the functions
directly, as they are called in Capstone, it crashes because the
actual functions are function pointers and are not called directly
in Capstone.  I'm not sure how important unifying the memory
management is, it's just that we may need to replace the standard
allocator since we want the superset disassembler to operate
outside the standard operations of a binary.