Implemented returning jmp instruction for encountered offsets
Now the superset disassembly algorithm should be working correctly,
although it behaves slightly differently from the previous
implementation.  In this case, instead of returning a sort of NULL
value when we try to disassemble an already encountered offset, I
instead decided to return a jmp instruction that jumps to itself.
This allows for a rewriter using this to rewrite that jmp to point
to the actual address (the first location we encountered that
offset) without using a special case.  The previous implementation
would just return None (in python), which required the code calling
the disassembler to distinguish these cases.

The algorithm also only inserts a jmp instruction when at least one
valid instruction was encountered starting from each offset.  This
prevents junk jmp instructions from being inserted, and simply uses
the jmp instructions to tie the different disassembled "layers"
(from different offsets) together.

The starting offsets are chosen by simply incrementing the starting
offset by one until an unencountered offset is found.  While this
is a natural way to approach this, it does have an impact on what
contiguous chunks of code are found.  Depending on the way
instructions overlap, this may result in many tiny sequences of
instructions at the end of the disassembly that all jump to longer
sequences near the start of the disassembly.  If one of these tiny
sequences is a valid target, this would add an extra unconditional
jump of overhead.  Whether this is a big deal, and what approaches
could be taken to try to find the "intended" sequences and merge
them, is a separate issue, but I do think it's important to
consider the implications of each implementation decision on
overhead.