Restructured disassembly code, added returning hlt on invalid inst
Now the returned disassembly terminates sequences that end in an
invalid instruction with hlt, and sequences that end in a
previously visited address with a jmp to itself.  I think I still
will need to differentiate these "special" instructions from real
instructions so they can be cleanly distinguished by anything
using this.

Since we likely don't need sequences that terminate in an invalid
instruction (past the last unconditional control flow instruction),
I need to make it easy for code generation that uses this to trim
away these bad sequences.  I still think it's fine and maybe even
desirable to return these suspect instructions, depending on what
kind of assumptions we are making about the binary, but under
normal conditions we will want to eliminate these when doing
rewriting for security reasons.  However, I think that step can be
done in the code generation phase, but I do need to provide a way
to identify the "special" instructions so that it's very clear what
instructions were in the original and which were added to mark
the ends of sequences.