Added compilation options for building 32-bit version of library
Now the disassembler can be optionally compiled for x86 instead of
x86-64.  This will be important if the disassembler needs to be
used on 32-bit systems.  For now, the way that the 32-bit version
of the Capstone library is handled will not be portable, because
the 32-bit version was specially named and copied into /usr/lib/
so that I could have both versions available at once.  This may
need to be done in a different way for a "release version", such
as by requiring the person compiling it to install their preferred
version of capstone first, and then compiling with a compile-time
flag/environment variable/option.