Fixed 64-bit support, fixed minor bug, general cleanup
I have now made the way 32-bit and 64-bit versions of the library
are built less inconsistent, and I added -Wall and -Wextra warning
flags to the Makefile.  This led to some cleanup in the test driver
files.  I also hadn't fixed the 64-bit driver to work with my
changes, so that works now.

The differences in the code I had noticed previously actually seem
to have been an improvement, as previously hlt instructions were
not appearing at the end of the code sequences where they should
have been.

However, when I tested with x64, I noticed that invalid
instructions were appearing in the disassembly, so I have now fixed
that so now a hlt is emitted upon encountering an invalid
instruction.  If we don't know what an instruction is, it's better
to make sure the program crashes right there to make it easier to
debug.  This may have been caused by udis86 having different
behavior to Capstone, as Capstone may have returned failure on
encountering an invalid instruction, while udis86 appears to return
success.