Fixed segfault caused by attempted read off end of mapping
While the underlying disassembler is able to determine when we have
reached the end of a code buffer, the way I was relying on this
caused an issue because before attempting to disassemble I checked
whether an instruction had been disassembled by referring to a
mapping for every address.  In certain cases, it was possible that
when the disassembly was at the very end of the code buffer,
attempting to read off the end of the mapping would cause a
segfault.  Now, I check whether the disassembled code sequence
reaches the end of the code buffer and return a special hlt
instruction if so.